Add some tap dance stuff, but don't turn it on until timing is figured out
The hold for the shift is not getting registered unless there is a very short tapping term.
But if there is a very short tapping term, the double and triple tap don't register.
Plus this may have to be set on a per key basis, becuase when the the tapping term is super short,
I also lose the Esc/Ctrl tap/hold ability on the CapsLock Key.